bank-statements-api/pyproject.toml
[project]
name = "bank-statements-api"
version = "0.1.0"
description = "API for Bank Statement Analyzer"
authors = [
    {name = "Pedro Sousa", email = "pedronsousa@gamil.com"}
]
requires-python = ">=3.9"
dependencies = [
    "jsonfinder>=0.4.2",
    "alembic>=1.10.3",
    "fastapi>=0.115.12",
    "google-genai>=1.9.0",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "pandas>=2.1.4",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.8",
    "sqlalchemy>=2.0.9",
    "uvicorn>=0.21.1",
    "openpyxl>=3.1.5",
    "authlib>=1.3.0",
    "pyjwt>=2.8.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "pytest-cov>=6.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 128
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 128

[tool.uv]
dev-dependencies = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "pytest-cov>=6.1.1",
]